refactor: move _list_repo_files and _repo_exists to hf_hub_utils module
Agent-Logs-Url: https://github.com/EuroEval/EuroEval/sessions/3de2ac2e-a2d9-46b1-afbb-20498690e000
