+++ src/euroeval/custom_dataset_configs.py
@@ -13,50 +13,12 @@ from types import ModuleType
 
 from huggingface_hub import HfApi
 
-from .caching_utils import cache_arguments
 from .data_models import DatasetConfig
+from .hf_hub_utils import _list_repo_files, _repo_exists
 from .logging_utils import log_once
 from .split_utils import get_repo_splits
 from .utils import get_hf_token
 from .yaml_config import load_yaml_config
-
-
-@cache_arguments("dataset_id", "revision")
-def _list_repo_files(hf_api: HfApi, dataset_id: str, revision: str) -> list[str]:
-    """Cached wrapper around ``HfApi.list_repo_files`` for dataset repos.
-
-    Args:
-        hf_api:
-            The Hugging Face API client.
-        dataset_id:
-            The dataset repo id.
-        revision:
-            The git revision to list files from.
-
-    Returns:
-        The list of file paths in the dataset repo at ``revision``.
-    """
-    return list(
-        hf_api.list_repo_files(
-            repo_id=dataset_id, repo_type="dataset", revision=revision
-        )
-    )
-
-
-@cache_arguments("dataset_id")
-def _repo_exists(hf_api: HfApi, dataset_id: str) -> bool:
-    """Cached wrapper around ``HfApi.repo_exists`` for dataset repos.
-
-    Args:
-        hf_api:
-            The Hugging Face API client.
-        dataset_id:
-            The dataset repo id.
-
-    Returns:
-        True if the dataset repo exists on the Hub, otherwise False.
-    """
-    return hf_api.repo_exists(repo_id=dataset_id, repo_type="dataset")
 
 
 def load_custom_datasets_module(custom_datasets_file: Path) -> ModuleType | None:
